+++ gems/ViewTrans.c
@@ -34,7 +34,7 @@ void BuildViewTransform( VRP, EP, UP, T )
 	 * Make vector  V  orthogonal to  N  and normalize  V
 	 */
 	V.x = UP.x - VRP.x; V.y = UP.y - VRP.y; V.z = UP.z - VRP.z;
-	dot = V3Dot(&V,&N);
+	dot = (float)V3Dot(&V,&N);
 	V.x -= dot * N.x; V.y -= dot * N.y; V.z -= dot * N.z;
 	V3Normalize(&V);
 
@@ -48,23 +48,23 @@ void BuildViewTransform( VRP, EP, UP, T )
 	 * Write the vectors U, V, and N as the first three rows of the
 	 *       first, second, and third columns of  T, respectively
 	 */
-	T[0][0] = U.x;		/* column 1 , vector U */
-	T[1][0] = U.y;
-	T[2][0] = U.z;
-	T[0][1] = V.x;		/* column 2 , vector V */
-	T[1][1] = V.y;
-	T[2][1] = V.z;
-	T[0][2] = N.x;		/* column 3 , vector N */
-	T[1][2] = N.y;
-	T[2][2] = N.z;
+	T[0][0] = (float)U.x;		/* column 1 , vector U */
+	T[1][0] = (float)U.y;
+	T[2][0] = (float)U.z;
+	T[0][1] = (float)V.x;		/* column 2 , vector V */
+	T[1][1] = (float)V.y;
+	T[2][1] = (float)V.z;
+	T[0][2] = (float)N.x;		/* column 3 , vector N */
+	T[1][2] = (float)N.y;
+	T[2][2] = (float)N.z;
 
 	/*
 	 * Compute the fourth row of  T  to include the translation of
 	 *       VRP  to the origin
 	 */
-	T[3][0] = - U.x * VRP.x - U.y * VRP.y - U.z * VRP.z;
-	T[3][1] = - V.x * VRP.x - V.y * VRP.y - V.z * VRP.z;
-	T[3][2] = - N.x * VRP.x - N.y * VRP.y - N.z * VRP.z;
+	T[3][0] = (float)(- U.x * VRP.x - U.y * VRP.y - U.z * VRP.z);
+	T[3][1] = (float)(- V.x * VRP.x - V.y * VRP.y - V.z * VRP.z);
+	T[3][2] = (float)(- N.x * VRP.x - N.y * VRP.y - N.z * VRP.z);
 
 	return;
 }
